protoc-gen-nrpc: handle output types matching the reply pattern
It a method output type has a single top-level 'oneof' named 'reply'
that contains two fields 'result' and 'error', then we generate a
slightly different code:

The server handler and client wrapper functions returns directly a
'result' field type instead of the output type itself.

The code around (or inside) these functions also changes to properly
build the output type from the result or extract the result from the
output type.

The 'result' and 'error' fields must be named 'Result' and 'Error' in
the go generated file for this patch to work.