Generates a client that regroups all the service clients
If a client program needs to use all the services of a package, it can become
tedious to setup and keep a hold on each service client.

This patch generates a package-level client that regroup all the service client in
a single struct, with some helpers to set all the timeouts or encoding.
It also holds the services parameters.