Cancel requests that are pending for too long
If a request is pending for more than the maxPendingDuration, the scheduler send a SERVERTOOBUSY
to the caller. This way, an overloaded server will always reply something to the client,
which reduces a lot the cases where it gets a timeout error.

The maxPendingDuration should be short enough so that this duration plus the handling
duration remains smaller than the client timeout.

A next version of this system could cancel a running handler if its too long and
send a reply 'server is too slow' to the client